+++ tests/integration/mod.rs
@@ -1,4 +1,5 @@
 mod creds;
 mod pub_properties;
 mod pub_sub;
+mod session;
 mod sub_options;